+++ public/locales/en/resumeProjects.json
@@ -4,26 +4,30 @@
   "projects": [
   {
     "title": "SmartStore",
-    "description": "Technology marketplace with shopping cart, payments simulator and registration forms for users and products.",
+    "description": "Technology marketplace where any user can sign up either to buy or sell products. Features a products catalog with search filters and shopping cart, a user control panel, and a payments simulator.",
     "website": "",
-    "imgUrl": "/imgs/projects/Smartstore1.webp"
+    "imgUrl": "/imgs/projects/Smartstore1.webp",
+    "technologies": []
   },
   {
     "title": "Pawsome Friends",
-    "description": "Website for adoption of abandoned pets. Includes a Home page, a pets catalog with search filters, and an admin control panel for pet shelters.",
+    "description": "Web app to promote the adoption of abandoned pets. Pet shelters can sign up to manage their registered pets, and any person who wants to adopt has to submit a form providing contact details and related info",
     "website": "https://pawsome-friends.netlify.app",
-    "imgUrl": "/imgs/projects/Pawsome1.webp"
+    "imgUrl": "/imgs/projects/Pawsome1.webp",
+    "technologies": []
   },
   {
     "title": "Issue Tracker",
     "description": "An app to enhance performance of HTTP requests, with caching, server state management and error handling, using Axios and TanStack Query libraries.",
     "website": "",
-    "imgUrl": "/imgs/projects/Issues1.webp"
+    "imgUrl": "/imgs/projects/Issues1.webp",
+    "technologies": []
   },
   {
     "title": "TicTapToe",
     "description": "Cool looking tic tac toe game developed from scratch, features a neumorphic design, scores bar autosave, shape and starting player switches, as well as local storage of scores.",
     "website": "https://tictaptoe.netlify.app",
-    "imgUrl": "/imgs/projects/Tictaptoe1.webp"
+    "imgUrl": "/imgs/projects/Tictaptoe1.webp",
+    "technologies": []
   }
 ]}
