+++ src/app/home.tsx
@@ -1,7 +1,6 @@
-import Main from "../components/homePage/Main";
+import Main from "../features/homePage/MainSection";
 
 export default function HomePage() {
-
   return (
     <main className="mt-6 w-full mx-auto">
       <Main></Main>
