+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Header from "../components/common/Header";
+import Header from "../components/Header";
 
 export default function RootLayout() {
   return (
