+++ src/features/homePage/MainSection.tsx
@@ -3,9 +3,11 @@ import { useTranslation } from "react-i18next";
 import ProjectCard from "./ProjectCard";
 import SkillSetCard from "./SkillSetCard";
 import JobCard from "./JobCard";
-import { Job, Project, SkillSet } from "../../types";
+import { Project } from "../projects/projectTypes";
+import { Job } from "../jobs/jobsTypes";
+import { SkillSet } from "../skills/skillTypes";
 
-export default function Main() {
+export default function MainSection() {
   const { t } = useTranslation(["resumeAbout", "resumeProjects", "resumeJobs", "resumeSkills"]);
   return (
     <Accordion className="max-w-full">
