+++ src/features/jobs/jobsTypes.ts
@@ -1,0 +1,10 @@
+export interface Job {
+  key: string;
+  role: string;
+  company: string;
+  date: string;
+  location: string;
+  url: string;
+  imgUrl: string;
+  description: string[];
+}
