+++ src/features/projects/projectTypes.ts
@@ -1,0 +1,10 @@
+import { Skill } from "../../types";
+
+
+export interface Project {
+    title: string;
+    description: string;
+    technologies: Skill[],
+    website: string;
+    imgUrl: string;
+  };
