+++ src/features/skills/skillTypes.ts
@@ -1,0 +1,9 @@
+export interface SkillSet {
+  type: string,
+  list: Skill[],
+}
+
+export interface Skill {
+  name: string,
+  imgUrl: string,
+}
